% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_outer.R
\name{plot.bma}
\alias{plot.bma}
\title{Plot Posterior Model Size and Model Probabilities}
\usage{
\method{plot}{bma}(x, ...)
}
\arguments{
\item{x}{an object of class 'bma'}

\item{\dots}{additional arguments for \code{\link{matplot}}}
}
\value{
combines the plotting functions \code{\link{plotModelsize}} and
\code{\link{plotConv}}
}
\description{
Produces a combined plot: upper row shows prior and posterior model size
distribution, lower row shows posterior model probabilities for the best
models
}
\note{
The upper plot shows the prior and posterior distribution of model
sizes (\code{\link{plotModelsize}}).\cr The lower plot is an indicator of
how well the bma object has converged (\code{\link{plotConv}}).
 and Paul
}
\examples{

data(datafls)
mm=bms(datafls,user.int=FALSE)

plot(mm)
}
\seealso{
\code{\link{plotModelsize}} and \code{\link{plotConv}}

Check \url{http://bms.zeugner.eu} for additional help.
}
\keyword{hplot}
