% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{class}
\name{zlm-class}
\alias{zlm-class}
\title{Class "zlm"}
\description{
A list holding output from the Bayesian Linar Model under Zellner's g prior
akin to class 'lm'
}
\section{Objects from the Class}{
 Objects can be created via calls to
\code{\link{zlm}}, but indirectly also via \code{\link{as.zlm}}.\cr
\code{\link{zlm}} estimates a Bayesian Linear Model under Zellner's g prior
- its output is very similar to objects of class \code{\link{lm}} (cf.
section 'Value')
}

\references{
\url{http://bms.zeugner.eu}
}
\seealso{
\code{\link{zlm}} and \code{\link{as.zlm}} for creating \code{zlm}
objects,\cr \code{\link{density.zlm}}, \code{\link{predict.zlm}} and
\code{\link{summary.zlm}} for other posterior results
}
\author{
Martin Feldkircher and Stefan Zeugner
}
\keyword{classes}
